Create Report Lambda Function

+++ src/main/java/edu/wpi/cs/yaml/demo/http/DeleteChoicesRequest.java
@@ -1,15 +1,15 @@
 package edu.wpi.cs.yaml.demo.http;
 
 public class DeleteChoicesRequest {
-	int howOld;
+	float howOld;
 	
-	public int returnHowOld() {return this.howOld;}
-	public void setHowOld(int howOld) {this.howOld = howOld;}
+	public float returnHowOld() {return this.howOld;}
+	public void setHowOld(float howOld) {this.howOld = howOld;}
 	
 	public DeleteChoicesRequest() {}
-	public DeleteChoicesRequest(int howOld) {this.howOld = howOld;}
+	public DeleteChoicesRequest(float howOld) {this.howOld = howOld;}
 	
 	public String toString() {
-		return "Requesting to delete all choices older than "+ Integer.toString(howOld) + " days";
+		return "Requesting to delete all choices older than "+ Float.toString(howOld) + " days";
 	}
 }
